Aggiunta di commenti ed eliminazione delle righe con valori nulli sulla colonna class
